Added conio.h with implementations of: gotoxy, clrscr
Added conio.h as part of winbgi to handle some console functions from the old Borland C++ conio.h. Like gotoxy, clrscr and CreateNewConsole and ReleaseConsole are also added to add or remove a console window when in graphical mode, and in this way easily load input data and display output or error messages when exiting graphical mode.